update to responsive text media query
tweak to media query as wasn't previoulsy accounting for landscape view
on phones with taller screens. for example, looking at a hotgraphic on
an iphone 5 with these responsive text classes, if you were to look at
it in landscape (which is wider that 520px device width small) you'd see
the desktop text and the mobile text would hide (so in this example,
you'd have hotgraphic instruction appearing on a narrative).